refactor: change auth logic
Change the auth logic to hide the password when it returns a user. Also, add some validations to make the auth proccess more seccure.
Add validation to allow superadmin to access all protected routes.
