+++ src/auth/auth.controller.ts
@@ -6,7 +6,7 @@ import { AuthGuard } from './guard/auth.guard';
 import { Request, request } from 'express';
 import { Roles } from './decorators/roles.decorators';
 import { RolesGuard } from './guard/roles.guard';
-import { Role } from './enums/rol.enum';
+import { Role } from '../common/enums/rol.enum';
 import { Auth } from './decorators/auth.decorators';
 
 // Define a new interface called RequestWithUser that extends the Request interface from Express. This interface will be used to define the user property in the request object.
