+++ src/auth/decorators/auth.decorators.ts
@@ -1,5 +1,5 @@
 import { UseGuards, applyDecorators } from "@nestjs/common";
-import { Role } from "../enums/rol.enum";
+import { Role } from "../../common/enums/rol.enum";
 import { Roles } from "./roles.decorators";
 import { AuthGuard } from "../guard/auth.guard";
 import { RolesGuard } from "../guard/roles.guard";
