+++ src/auth/guard/roles.guard.ts
@@ -1,7 +1,7 @@
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { ROLES_KEY } from '../decorators/roles.decorators';
-import { Role } from '../enums/rol.enum';
+import { Role } from '../../common/enums/rol.enum';
 
 /**
  * Create a new guard called RolesGuard that implements the CanActivate interface. This guard will check if the user has the required role to access the route.
@@ -35,6 +35,10 @@ export class RolesGuard implements CanActivate {
 
     const { user } = context.switchToHttp().getRequest();
     const userRole = user.role;
+
+    if (userRole === Role.SUPER_ADMIN) {
+      return true;
+    }
 
     return roles.includes(userRole);
   }
