+++ src/users/users.service.ts
@@ -22,8 +22,15 @@ export class UsersService {
     return this.usersRepository.findOneBy({ email });
   }
 
+  findByEmailWithPassword(email: string) {
+    return this.usersRepository.findOne({ 
+      where: { email }, 
+      select: ['id', 'name', 'email', 'password', 'role'] 
+    });
+  }
+
   findAll() {
-    return `This action returns all users`;
+    return this.usersRepository.find();
   }
 
   findOne(id: number) {
